Fix print formatting for MINGW build

+++ src/lib/public/quadtree.h
@@ -673,8 +673,8 @@
         while(sib) {                                                                            \
             qt_node(name) *sib_node = mp_##name##_entry(&qt->node_pool, sib);                   \
             if(sib_node->depth != node->depth || sib_node->parent != node->parent) {            \
-                printf("Mismatching sibling depths/parents for node (%f, %f) [%zu] and its' sibling (%f, %f) [%zu]\n", \
-                    node->x, node->y, node->depth, sib_node->x, sib_node->y, sib_node->depth);  \
+                printf("Mismatching sibling depths/parents for node (%f, %f) [%d] and its' sibling (%f, %f) [%d]\n", \
+                    node->x, node->y, (int)node->depth, sib_node->x, sib_node->y, (int)sib_node->depth); \
                 qt_##name##_print(qt);                                                          \
                 fflush(stdout);                                                                 \
             }                                                                                   \
@@ -686,8 +686,8 @@
         if(node->parent) {                                                                      \
             qt_node(name) *parent = mp_##name##_entry(&qt->node_pool, node->parent);            \
             if(node->depth - parent->depth != 1) {                                              \
-                printf("Mismatching depths for node (%f, %f) [%zu] and its' parent (%f, %f) [%zu]\n", \
-                    node->x, node->y, node->depth, parent->x, parent->y, parent->depth);        \
+                printf("Mismatching depths for node (%f, %f) [%d] and its' parent (%f, %f) [%d]\n", \
+                    node->x, node->y, (int)node->depth, parent->x, parent->y, (int)parent->depth); \
                 qt_##name##_print(qt);                                                          \
                 fflush(stdout);                                                                 \
             }                                                                                   \
